reorganizing weight analysis code
